I choose csv specifically, because they are easy to read and import into other softwares (like spreadsheet app).
The idea is to enable version control of the project planning sheet.
Although there are number of tools to enable project planning, but
all of them are very heavy and doesnt allow data to be imported into
version control. You could let the project data alongside the source
or alternatively create another repository just for the project
and time sheets.

You will have to wait a bit longer before this project does
anything useful for the general audience.